requirements.txt
Flask==3.1.2
Jinja2==3.1.2
click==8.1.6
Flask-WTF==1.1.1
requests==2.32.1
pymysql==1.1.1
Flask-SQLAlchemy==3.1.1
numpy
pandas
matplotlib
seaborn
scikit-learn
python-dotenv
gunicorn

